requirements.txt
pyTelegramBotAPI
mercadopago
python-dotenv
